+++ src/analyzer/error/Error.hpp
@@ -53,7 +53,7 @@ public:
 	std::vector<std::string> parameters;
 	std::string underline_code;
 
-	Error(Type type, File* file, int line, int character);
+	Error(Type type, File* file, size_t line, size_t character);
 	Error(Type type, Token* token, std::vector<std::string> parameters);
 	Error(Type type, Location location, Location focus);
 	Error(Type type, Location location, Location focus, std::vector<std::string> parameters);
